Update gpg path
The gpg suite is installed on /usr/local/bin, and the plugin doesn't work without it. This is just to test if it works and I like the plugin. If I do, I'll probably try to submit a PR to add that as an option to pass to the plugin.